+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/components/GridUtils.groovy
@@ -1,0 +1,10 @@
+package life.qbic.portal.sampletracking.components
+
+/**
+ * A helper class with static utility functions for Vaadin Grids.
+ *
+ * @since 1.0.0
+ */
+class GridUtils {
+    final static int TITLE_COLUMN_WIDTH = 400
+}